+++ era_customer_success/models/__init__.py
@@ -22,3 +22,5 @@ from . import cs_voc_insight
 from . import account_return
 from . import self_service_help
 from . import cs_user_guide
+from . import cs_portfolio_share
+from . import google_sheet_sync